Downstream improvements to English text
I have revised many places again. In the Chinese file, I refer to the latest version of the English text, and amend the English text to be smoother by the way, or add an explanation of the actual situation, so that users can better understand how to use each command or function.